Move devices logic under separate directory
Next stage in refactoring device management.
Move all related logic under a dedicated devices directory.
